+++ src/assignments/maintenance/index.ts
@@ -1,0 +1,2 @@
+export { SupersededCleanupService } from './superseded-cleanup.service';
+export { OrphanCleanupService } from './orphan-cleanup.service';
